requirements.txt
Flask==3.1.3
joblib==1.5.3
pandas==2.3.2
scikit-learn==1.8.0
google-genai==1.68.0
python-dotenv==1.2.2
gunicorn